Add files via upload
I did some Data Preprocessing and Data reduction , but I found there is no omitted value for each row, 
so maybe this process would be rather explore , understand each variables, and relation between variables. 
And Maybe We will use lm , or knn(for numerical) to predict market cap, and also during lm, maybe we can use search(exhaustive 
and local) to find if there is a possibility of data reduction.