Add files via upload
It's  our midle report for submit! and 
I writed some parts of this report. But not yet completed.. 
If you guys read the report ,  and could you append the rest of parts, or give me a feedback?? 
then I 'll edit and change the expression of sentences ,, and so on !